docker: Add compose.dev.yaml for local development, watch mode

+++ website/bootstrap.sh
@@ -9,18 +9,19 @@ get_public_meilisearch_key() {
 export NEXT_PUBLIC_MEILISEARCH_KEY=$(get_public_meilisearch_key)
 echo the public meilisearch key is: $NEXT_PUBLIC_MEILISEARCH_KEY
 
-# Must be done when Strapi is already up and running.
-# (That's why it's not in a Dockerfile).
-# Also, the public key gets embedded at this stage.
-npm run build
 
-# Purge the NGINX cache; otherwise, there will be a mismatch between the
-# `nginx` cache and the `next.js` server, negatively impacting the SPA
-# experience. However, nothing bad will happen: the site will still be usable.
-rsync -a /opt/website/builds/current/static/* /nextjs_static
-rm -fr /opt/website/builds/current/static
-ln -s /nextjs_static /opt/website/builds/current/static
-rm -fr /nginx_cache/*
+if [ "$1" = "--dev" ]; then
+  echo "Running in development mode"
+  npm run dev
+else
+  # Must be done when Strapi is already up and running. That's why it's not in
+  # a Dockerfile. Also, the meilisearch public key (see above) gets embedded at
+  # this stage.
+  npm run build
 
+  rsync -a /opt/website/builds/current/static/* /nextjs_static
+  rm -fr /opt/website/builds/current/static
+  ln -s /nextjs_static /opt/website/builds/current/static
 
-node /opt/website/builds/current/standalone/server.js
+  node /opt/website/builds/current/standalone/server.js
+fi
